+++ src/main/java/com/nothome/delta/DiffWriter.java
@@ -32,7 +32,7 @@ import java.io.IOException;
  * @author  torgeir
  */
 public interface DiffWriter {
-    public void addCopy(int offset, int length) throws IOException;
+    public void addCopy(long offset, int length) throws IOException;
 	public void addData(byte b[], int offset, int length) throws IOException;
     public void flush() throws IOException;
     public void close() throws IOException;
